+++ app/background-process.js
@@ -9,6 +9,7 @@ import env from './env'
 import * as windows from './background-process/windows'
 import buildMenu from './background-process/window-menu'
 import * as sitedata from './background-process/sitedata'
+import * as bookmarks from './background-process/bookmarks'
 
 import * as ipfsNetwork from './background-process/networks/ipfs'
 
@@ -26,7 +27,10 @@ app.on('ready', function () {
   // ui
   Menu.setApplicationMenu(Menu.buildFromTemplate(buildMenu(env)));
   windows.setup()
+
+  // databases
   sitedata.setup()
+  bookmarks.setup()
 
   // networks
   ipfsNetwork.setup()