+++ app/builtin-pages/start.js
@@ -1,5 +1,3 @@
-import * as bookmarksData from './data/bookmarks'
 import * as bookmarksUI from './ui/bookmarks'
 
-bookmarksData.setup()
 bookmarksUI.setup()